% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{asestio}
\alias{asestio}
\alias{asestio-package}
\title{asestio: An R Data I/O & Transformation Toolkit for ASES}
\description{
An analytics toolkit comprising of a series of work functions on data I/O and transformation.
The toolkit has a lot of featured funcionalities specifically designed for ASES HMS team.
}
\details{
The toolkit comprises of several types of functions: \describe{
\item{File conversion}{\code{\link{conv_pdfs}()}, \code{\link{xls2xlsx}()},
  \code{\link{doc2docx}()}}
\item{Merge datasets}{\code{\link{bind_pdfs}()}, \code{\link{bind_txts}()},
  \code{\link{bind_excels}()}}
\item{Merge data rows/columns}{\code{\link{bind_cols}()}, \code{\link{bind_rows}()}}
\item{Other}{\code{\link{rehead_htmltbl}()}, \code{\link{simp_summ}()},
  \code{\link{find_var_row}()}}
}
}
\seealso{
\pkg{\link{recharts}}
}
\author{
\strong{Maintainer}: Yiying Wang, \email{wangy@aetna.com}
}
\keyword{internal}
