% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_col.R
\name{bindColumns}
\alias{bindColumns}
\alias{bind_columns}
\title{Combine Similar Columns Into One}
\usage{
bindColumns(mapping.file = paste0(toolkit, "BindFieldNames.xlsx"),
  toolkit = getOption("toolkit.dir"), sheet = 1,
  output.format = c("xlsx", "csv"))

bind_columns(mapping.file = paste0(toolkit, "BindFieldNames.xlsx"),
  toolkit = getOption("toolkit.dir"), sheet = 1,
  output.format = c("xlsx", "csv"))
}
\arguments{
\item{mapping.file}{Name of the field name mapping dictionary. Default
'BindFieldNames.xlsx' in \code{toolkit} directory.}

\item{toolkit}{Directory of the dictionary.}

\item{sheet}{the sheet (number or name character) of the target workbook to
read in. Default 1.}

\item{output.format}{csv or xlsx. Default xlsx.}
}
\value{
A list. \itemize{
 \item \code{Output Results} shows the path where bind_output.xlsx is;
 \item \code{Dropped columns} shows the variables dropped during the combination.
}
}
\description{
Combine similar columns into one column within one spreadsheet.
}
\examples{
\dontrun{
bindColumns()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
