% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_excel.R
\name{bindExcels}
\alias{bindExcels}
\alias{bind_excels}
\title{Combine Excel Files}
\usage{
bindExcels(sheet.num = 1, files = NULL, output.format = c("xlsx",
  "csv"))

bind_excels(sheet.num = 1, files = NULL, output.format = c("xlsx",
  "csv"))
}
\arguments{
\item{sheet.num}{Sheet number where data is stored. If null,
then \pkg{\link{dplyr}}::\code{\link[dplyr]{bind_rows}} all the sheets.}

\item{files}{Designate the files yourself, or manually select them.}

\item{output.format}{csv or xlsx. Default xlsx.}
}
\value{
A list containing a data frame listing the files and sheets combined.
}
\description{
Combine several excel files into one.
}
\examples{
\dontrun{
bindExcels()
}
}
\seealso{
\code{\link[dplyr]{bind_rows}} in package \pkg{\link{dplyr}}.
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
