% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pdf.R
\name{bindPDFs}
\alias{bindPDFs}
\alias{bind_pdfs}
\title{Convert PDFs to TXTs And Combine Them to A Data.frame}
\usage{
bindPDFs(mode = 1, converter = c("pdftotext", "pdftohtml"),
  converter.path = NULL, rule.file = paste0(path,
  "IndRptToTable(RegExp).xlsx"), path = getOption("toolkit.dir"))

bind_pdfs(mode = 1, converter = c("pdftotext", "pdftohtml"),
  converter.path = NULL, rule.file = paste0(path,
  "IndRptToTable(RegExp).xlsx"), path = getOption("toolkit.dir"))
}
\arguments{
\item{mode}{integer, Default 1L. \describe{
\item{1L}{as is (-raw)} \item{2L}{with layout (-layout)}
\item{3L}{no breaks (-nopgbrk)} \item{4L}{without format ()}
\item{5L}{as table (-table)} \item{6L}{simplified (-simple)}
}}

\item{converter}{The converter program to use, either "pdftotext" or "pdftohtml"}

\item{converter.path}{Path to the converter program. If NULL, make sure the
xpdf folder is under R.home(). E.g., \code{paste0(R.home(),
paste0("/xpdfbin-win-4.00/bin64/pdftotext.exe"))}}

\item{rule.file}{Full path of the conversion dictionary file. Default
\code{paste(path, <dict name>)}.}

\item{path}{Path of the directory to the dictionary file.}
}
\value{
Nothing
}
\description{
This is a function combining \code{\link{convPDFs}} and \code{\link{bindTXTs}}.
You can directly select PDFs and then convert them to txts, based on which you
can then combine them to a data.frame with a mapping rule.file.
}
\examples{
\dontrun{
bindPDFs()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
