% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_row.R
\name{bindRows}
\alias{bindRows}
\alias{bind_rows}
\title{Combine Rows Into One}
\usage{
bindRows(df = NULL, output.format = c("xlsx", "csv"))

bind_rows(df = NULL, output.format = c("xlsx", "csv"))
}
\arguments{
\item{df}{You can assign a data.frame yourself. If null, you will manually
select a spreadsheet and extract the data.}

\item{output.format}{xlsx or csv. Default xlsx.}
}
\value{
Nothing
}
\description{
Combine rows into one which are actually sole record.
}
\examples{
\dontrun{
bindRows()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
