% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_docx.R
\name{convDoc2Docx}
\alias{convDoc2Docx}
\alias{doc2docx}
\title{Convert doc to docx}
\usage{
convDoc2Docx(files = NULL, delete.original = FALSE,
  use.parallel = TRUE)

doc2docx(files = NULL, delete.original = FALSE, use.parallel = TRUE)
}
\arguments{
\item{files}{character vector of dic files. Default NULL (A GUI wizard is
then called).}

\item{delete.original}{Logical. If TRUE, then the original doc files will be
deleted after the conversion. Default FALSE.}

\item{use.parallel}{Logical, whether or not to use parallel mode when 
there are more than 50 files. Default TRUE.}
}
\value{
Nothing
}
\description{
Convert doc to docx in batch manner
}
\examples{
\dontrun{
convDoc2Docx()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
