% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_xlsx.R
\name{convXls2Xlsx}
\alias{convXls2Xlsx}
\alias{xls2xlsx}
\title{Convert xls to xlsx}
\usage{
convXls2Xlsx(files = NULL, delete.original = FALSE,
  use.parallel = TRUE)

xls2xlsx(files = NULL, delete.original = FALSE, use.parallel = TRUE)
}
\arguments{
\item{files}{character vector of xls files. Default NULL (A GUI wizard is
then called).}

\item{delete.original}{Logical. If TRUE, then the original xls files will be
deleted after the conversion.}

\item{use.parallel}{Logical, if use parallel mode when there are more than
50 files. Default TRUE.}
}
\value{
Nothing
}
\description{
Convert xls to xlsx in batch manner
}
\examples{
\dontrun{
convXls2Xlsx()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
