% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_var_row.R
\name{findVarRow}
\alias{findVarRow}
\alias{find_var_row}
\title{Find the Variable Row}
\usage{
findVarRow(files = NULL)

find_var_row(files = NULL)
}
\arguments{
\item{files}{You can manually assign the files. If null, you will select them
using a wizard.}
}
\value{
A list \describe{
\item{rowindex}{row index}
\item{vars}{col names (table head)}
\item{checkrows}{a data.frame containing pct_valid_val, pct_unique_val, pct_complete_val}
}
}
\description{
Find the variable row and return a list of the index, row names and a data
frame containing pct_valid_val, pct_unique_val, pct_complete_val.
It makes use of the private function \pkg{aseshms}:::\code{checkTopRows}.
}
\examples{
\dontrun{
findVarRow()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
