% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_rpt_summ.R
\name{simpSumm}
\alias{simpSumm}
\alias{simp_summ}
\title{Simplify Checkup Data Sumamry}
\usage{
simpSumm(summ.col, file = NULL, sheet = 1, provider = c("Ciming",
  "Meinian", "Ikang"), ...)

simp_summ(summ.col, file = NULL, sheet = 1, provider = c("Ciming",
  "Meinian", "Ikang"), ...)
}
\arguments{
\item{summ.col}{The index or name of the summary column}

\item{file}{The checkup raw dataset in Excel spreadsheet format}

\item{sheet}{The index or name of the sheet. Default 1.}

\item{provider}{A string. Currently support "Ciming", "Meinian" and "Ikang"}
}
\value{
A xlsx file under the same directory, containing modified sumamry column
}
\description{
Simplify the summary column of checkup raw dataset.
}
\details{
There are 3 additional arguments: \cr
\describe{
  \item{beginner}{The beginner RegEx expression of a sentence.}
  \item{ender}{The ending RegEx expression of a sentence.}
  \item{special.chr}{Invoke the beginner/ender structure and keep the special.chr.}
}
}
\examples{
\dontrun{
simpSum("C", provider="Ikang")

##
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
